requirements.txt
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.0


